% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbox.R
\name{plotbox}
\alias{plotbox}
\title{Boxplot for one variable}
\usage{
plotbox(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Boxplot for one variable
}
\examples{
library(fsql)
rnorm(100) \%>\% plotbox
}
