% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdens.R
\name{plotdens}
\alias{plotdens}
\title{Quick density plot}
\usage{
plotdens(d, x, fill = "#69b3a2", color = "#e9ecef", alpha = 0.8)
}
\arguments{
\item{d}{data.frame}

\item{x}{variable whose density should be displayed}

\item{fill}{for the density curve aspect}

\item{color}{for the density curve aspect}

\item{alpha}{for the density curve aspect}
}
\value{

}
\description{
Quick density plot
}
\examples{
library(fsql)
d <- data.frame(x = rnorm(mean = 0,sd = 1,10000))
d \%>\% plotdens(x)
}
