% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plotfun}
\alias{plotfun}
\title{function plot}
\usage{
plotfun(f, minval, maxval, n = 1000)
}
\arguments{
\item{f}{one-argument function}

\item{minval}{xmin}

\item{maxval}{xmax}

\item{n}{steps number}
}
\value{
graph of f
}
\description{
function plot
}
\examples{
plotfun(\(x) sin(x^2), 0,6)
}
