% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsc.R
\name{plotsc}
\alias{plotsc}
\title{quick scatter plot}
\usage{
plotsc(d, x, y)
}
\arguments{
\item{y}{}
}
\value{
the scatter ggplot
}
\description{
quick scatter plot
}
\examples{
library(fsql)
d <- data.frame(x = c(1,2,3), y = c(5,4,6))
d \%>\% plotsc(x,y)
}
