% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotts.R
\name{plotts}
\alias{plotts}
\title{Plot several columns as functions of the column x}
\usage{
plotts(
  d,
  ...,
  x,
  freq = NULL,
  start = NULL,
  agg = function(u) mean(u, na.rm = T),
  multi = F,
  layout = NULL
)
}
\arguments{
\item{d}{data.frame}

\item{...}{the y-axis columns}

\item{x}{the x-axis column}

\item{freq}{frequency for resampling (see help for sampleby for details)}

\item{start}{start time for resampling}

\item{agg}{aggregate function for resampling}

\item{multi}{multiplot? boolean}

\item{layout}{ignored if multi=False. Vector indicating where goes what. Example: c(1,3,2) means three subcharts, the first one hwith the first curve, second one with the next three curves and the last one with the last two curves}
}
\value{
ggplot object if multi=F, nothing otherwise
}
\description{
Plot several columns as functions of the column x
}
\examples{

library(fsql)
d <- data.frame(
date=seq(as.Date("2001-01-01"), as.Date("2001-10-01"), by="month"),
ticker=sample(c("1 HK", "2 HK"), 10, TRUE),
ret1=rnorm(10),
ret2=rnorm(10),
ret3=rnorm(10)
) \%>\%
update(price1 = cumsum(ret1), price2 = cumsum(ret2), price3 = cumsum(ret3))


d \%>\% plotts(price1,price2,x=date)

d \%>\% plotts(price1,price2,x=date, freq="2M", agg = last)

d \%>\% plotts(price1, price2, price3, x=date, multi=T)

d \%>\% plotts(price1, price2, price3, x=date, multi=T, layout=c(2,1))
}
