% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvec.R
\name{plotvec}
\alias{plotvec}
\title{quick plot of a vector}
\usage{
plotvec(v)
}
\arguments{
\item{v}{vector}
}
\value{
the associated plot
}
\description{
quick plot of a vector
}
\examples{

library(fsql)
1:100 \%.>\% sqrt \%>\% plotvec
}
